cleanup after sublime3

+++ sublime3/install.sh
@@ -12,5 +12,11 @@ else
     fail "unsupported operating system"
 fi
 
-mv "$sublime_dir/User" "$sublime_dir/User.backup"
-ln -s ~/.dotfiles/sublime3/User "$sublime_dir"
+if ! [ -L "$sublime_dir/User" ]
+then
+    mv "$sublime_dir/User" "$sublime_dir/User.backup"
+    ln -s ~/.dotfiles/sublime3/User "$sublime_dir"
+    success "sublime3 installed"
+else
+    info "sublime3 already installed"
+fi
